pyproject.toml
[project]
name = "mural_digital"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"  # Debian Trixie comes with Python 3.13
dependencies = [
    "pillow>=12.0.0",
    "ruamel-yaml>=0.18.16",
]

[tool.ruff]
line-length = 120
